# Conflicts:
#   BE/algoarium/src/main/java/com/ssafy/algoarium/User/UserService.java
#   BE/algoarium/src/main/resources/application.properties

+++ FE/Algo/app/src/main/java/com/d204/algo/data/RankingRepositoryImpl.kt
@@ -24,7 +24,7 @@ class RankingRepositoryImpl @Inject constructor(
         emit(handleApi { dataSourceFactory.getDataStore(isRemote).getRankingsByTier(tier) })
     }
 
-    override suspend fun getRanking(userId: Int): Flow<NetworkResult<Ranking>> = flow {
+    override suspend fun getRanking(userId: Long): Flow<NetworkResult<Ranking>> = flow {
         emit(handleApi { dataSourceFactory.getRemoteDataSource().getRanking(userId) })
     }
 }
